cart_provider 다 연동되게 해놨고 price type 변경(int->double)!

+++ lib/pages/item_register.dart
@@ -257,7 +257,7 @@ class _ItemRegister extends State<ItemRegister> {
     final newItem = Item(
       id: id,
       name: nameController.text,
-      price: int.parse(priceController.text.replaceAll(",", "")),
+      price: double.parse(priceController.text.replaceAll(",", "")),
       imagePath: _image!.path,
       description: descriptionController.text,
     );
